Remove checkout page and redirect to Stripe from summary
