+++ too-rich-to-care-frontend/src/App.tsx
@@ -2,7 +2,6 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import SelectBillionaire from './pages/SelectBillionaire';
 import SpendMoney from './pages/SpendMoney';
 import Summary from './pages/Summary';
-import CheckoutReview from './pages/CheckoutReview';
 import PaymentSuccess from './pages/PaymentSuccess';
 import PaymentError from './pages/PaymentError';
 
@@ -12,7 +11,6 @@ function App() {
       <Routes>
         <Route path="/" element={<SelectBillionaire />} />
         <Route path="/spend" element={<SpendMoney />} />
-        <Route path="/checkout" element={<CheckoutReview />} />
         <Route path="/payment-success" element={<PaymentSuccess />} />
         <Route path="/payment-error" element={<PaymentError />} />
         <Route path="/summary" element={<Summary />} />
